\name{College}
\alias{College}
\docType{data}
\title{
Data for College Level Examination Program and the College Qualification Test
}
\description{
Scores obtained from 87 college students on the College Level Examination Program and the College Qualification Test.
}
\usage{College}
\format{
  A data frame with 87 observations and the following 3 variables.
  \describe{
    \item{\code{Science}}{Science (CQT) - numerical vector}
    \item{\code{Social}}{Social science and history (CLEP) - numerical vector}
    \item{\code{Verbal}}{Verbal (CQT) - numerical vector}
  }
}
\source{
Johnson, R.A., Wichern, D.W. (2002) Applied Multivariate Statistical Analysis. Prentice Hall.
}
