\name{mvdalab-package}
\alias{mvdalab-package}
\alias{mvdalab}
\docType{package}
\title{
Multivariate Data Analysis Laboratory (mvdalab)
}
\description{
Implementation of latent variables methods. The focus is on explorative anlaysis
using dimensionality reduction methods, such as Principal Component Analysis (PCA), 
and on multivariate regression based on Partial Least Squares regression (PLS). 
PLS analyses are supported by embedded bootstrapping and variable selection procedures.
}
\details{
\tabular{ll}{
Package: \tab mvdalab\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-10\cr
License: \tab GPL-3\cr

}
}
\author{Nelson Lee Afanador (\email{nelson.afanador@mvdalab.com}), Thanh Tran (\email{thanh.tran@mvdalab.com}), Lionel Blanchet (\email{lionel.blanchet@mvdalab.com})

Maintainer: Nelson Lee Afanador (\email{nelson.afanador@mvdalab.com})
}




