\name{Penta}
\alias{Penta}
\docType{data}
\title{
Penta data set
}
\description{
This data is obtained from drug discovery and includes measurements pertaining to size, lipophilicity, and polarity at various sites on a molecule. 
}
\usage{Penta}
\format{
  A data frame with 30 observations and the following 17 variables.
  \describe{
    \item{\code{Obs.Name}}{Categorical ID Variable}
    \item{\code{S1}}{numeric predictor vector}
    \item{\code{L1}}{numeric predictor vector}
    \item{\code{P1}}{numeric predictor vector}
    \item{\code{S2}}{numeric predictor vector}
    \item{\code{L2}}{numeric predictor vector}
    \item{\code{P2}}{numeric predictor vector}
    \item{\code{S3}}{numeric predictor vector}
    \item{\code{L3}}{numeric predictor vector}
    \item{\code{P3}}{numeric predictor vector}
    \item{\code{S4}}{numeric predictor vector}
    \item{\code{L4}}{numeric predictor vector}
    \item{\code{P4}}{numeric predictor vector}
    \item{\code{S5}}{numeric predictor vector}
    \item{\code{L5}}{numeric predictor vector}
    \item{\code{P5}}{numeric predictor vector}
    \item{\code{log.RAI}}{numeric response vector}
  }
}

\source{
Umetrics, Inc. (1995), Multivariate Analysis (3-day course), Winchester, MA.

SAS/STAT(R) 9.22 User's Guide, "The PLS Procedure".
}
