\name{plusMinusDat}
\alias{plusMinusDat}
\docType{data}
\title{
plusMinusDat data set
}
\description{
A simulated dataset for demonstrating the performance of a \code{plusminusFit} analysis. 
}
\usage{plusMinusDat}
\format{
  A data frame with 201 observations, 200 input variables (X) and one response variable (Y).
}

\source{
Richard Baumgartner (\email{richard_baumgartner@merck.com})
}
