\name{Wang_Chen}
\alias{Wang_Chen}
\docType{data}
\title{
Bivariate process data.
}
\description{
Twenty-five observations where 'H' represents brinell hardness and 'S' represents tensile strength.
}
\usage{Wang_Chen}
\format{
  A data frame with 25 observations and the following 2 variables.
  \describe{
    \item{\code{H}}{brinell hardness}
    \item{\code{S}}{tensile strength}
  }
}
\source{
Wang F, Chen J (1998). "Capability index using principal components analysis." Quality Engineering, 11, 21-27.
}
