\name{Wang_Chen_Sim}
\alias{Wang_Chen_Sim}
\docType{data}
\title{
Simulated process data from a plastics manufacturer.
}
\description{
Fifty observations where 'D' represents depth, 'L' represents length, and 'W' represents width.
}
\usage{Wang_Chen_Sim}
\format{
  A simulated data frame with 50 observations and the following 3 variables.
  \describe{
    \item{\code{D}}{depth}
    \item{\code{L}}{length}
    \item{\code{W}}{width}
  }
}
\source{
Data simulated by Nelson Lee Afanador from average and covariance estimates provided in Wang F, Chen J (1998). "Capability index using principal components analysis." Quality Engineering, 11, 21-27.
}

